moved game Cream Pie to the games package
