+++ gazeplay/src/main/java/net/gazeplay/games/creampie/CreampieStats.java
@@ -1,8 +1,9 @@
-package utils.games.stats;
+package net.gazeplay.games.creampie;
 
 import javafx.scene.Scene;
+import utils.games.stats.ShootGamesStats;
 
-public class CreampieStats extends ShootGamesStats{
+public class CreampieStats extends ShootGamesStats {
 
     public CreampieStats(Scene scene) {
         super(scene);
